+++ srcs/convert/cmd_args.c
@@ -6,15 +6,15 @@
 /*   By: jnakahod <jnakahod@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/27 16:54:14 by jnakahod          #+#    #+#             */
-/*   Updated: 2021/10/02 19:15:23 by jnakahod         ###   ########.fr       */
+/*   Updated: 2021/10/19 15:28:31 by jnakahod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <mini_shell.h>
 
-static int		counst_args(t_token *args)
+static int	counst_args(t_token *args)
 {
-	t_token *tmp;
+	t_token	*tmp;
 	int		res;
 
 	tmp = args;
@@ -30,9 +30,9 @@ static int		counst_args(t_token *args)
 const char	**create_cmd_args(t_token *args)
 {
 	const char	**res;
-	int		i;
-	int		ac;
-	t_token	*arg;
+	int			i;
+	int			ac;
+	t_token		*arg;
 
 	ac = counst_args(args);
 	res = (const char **)malloc(sizeof(char *) * (ac + 1));
@@ -40,7 +40,7 @@ const char	**create_cmd_args(t_token *args)
 		return (NULL);
 	arg = args;
 	i = 0;
-	while(arg)
+	while (arg)
 	{
 		res[i] = ft_strdup(arg->token);
 		arg = arg->next;
